feat: rename namespace from `res` to `result`
