+++ include/result/ok.hpp
@@ -1,15 +1,15 @@
 #pragma once
 
-namespace res {
+namespace result {
 
 /** A type that represents an ok status (success).
  * Use this type on res::Result<> to set the status of that data to be ok.
  *
  * @code
- * res::Result<> result = res::Ok();
- * assert(result.is_ok());
+ * result::Result<> res = result::Ok();
+ * assert(res.is_ok());
  * @endcode
  */
 struct Ok {};
 
-}  // namespace res
+}  // namespace result
